Fix a potential bug with PME and Amoeba on HIP
Some PME kernels on HIP are launched by executeKernelFlat(),
but the same sources are used in Amoeba which launches kernels using
execute().
It's possible that with small numThreadBlocks and large systems these
kernels are not able to process all required data.
USE_FLAT_KERNELS is used for kernels launched by executeKernelFlat().